- simplify cell events and usage
