+++ src/dash-table/derived/cell/contents.tsx
@@ -1,8 +1,6 @@
 import * as R from 'ramda';
 import React from 'react';
 
-import { memoizeOneFactory } from 'core/memoizer';
-
 import {
     ActiveCell,
     Data,
@@ -15,16 +13,15 @@ import {
     Presentation
 } from 'dash-table/components/Table/props';
 import CellInput from 'dash-table/components/CellInput';
-import derivedCellEventHandlerProps from 'dash-table/derived/cell/eventHandlerProps';
+import derivedCellEventHandlerProps, { Handler } from 'dash-table/derived/cell/eventHandlerProps';
 import isActiveCell from 'dash-table/derived/cell/isActive';
 import isCellEditable from './isEditable';
 import CellLabel from 'dash-table/components/CellLabel';
 import CellDropdown from 'dash-table/components/CellDropdown';
+import { memoizeOne } from 'core/memoizer';
 
 const mapData = R.addIndex<Datum, JSX.Element[]>(R.map);
 const mapRow = R.addIndex<IVisibleColumn, JSX.Element>(R.map);
-
-const cellEventHandlerProps = derivedCellEventHandlerProps();
 
 enum CellType {
     Dropdown,
@@ -48,69 +45,76 @@ function getCellType(
     }
 }
 
-const getter = (
-    activeCell: ActiveCell,
-    columns: VisibleColumns,
-    data: Data,
-    offset: IViewportOffset,
-    editable: boolean,
-    isFocused: boolean,
-    dropdowns: (DropdownValues | undefined)[][],
-    propsFn: () => ICellFactoryProps
-): JSX.Element[][] => mapData(
-    (datum, rowIndex) => mapRow(
-        (column, columnIndex) => {
-            const active = isActiveCell(activeCell, rowIndex + offset.rows, columnIndex + offset.columns);
+export default (propsFn: () => ICellFactoryProps) => new Contents(propsFn).get;
 
-            const dropdown = dropdowns[rowIndex][columnIndex];
-            const handlers = cellEventHandlerProps(propsFn)(rowIndex, columnIndex);
+class Contents {
+    constructor(
+        propsFn: () => ICellFactoryProps,
+        private readonly handlers = derivedCellEventHandlerProps(propsFn)
+    ) {
 
-            const isEditable = isCellEditable(editable, column.editable);
+    }
 
-            const className = [
-                ...(active ? ['input-active'] : []),
-                ...(isFocused ? ['focused'] : ['unfocused']),
-                ...['dash-cell-value']
-            ].join(' ');
+    get = memoizeOne((
+        activeCell: ActiveCell,
+        columns: VisibleColumns,
+        data: Data,
+        offset: IViewportOffset,
+        editable: boolean,
+        isFocused: boolean,
+        dropdowns: (DropdownValues | undefined)[][]
+    ): JSX.Element[][] => mapData(
+        (datum, rowIndex) => mapRow(
+            (column, columnIndex) => {
+                const active = isActiveCell(activeCell, rowIndex + offset.rows, columnIndex + offset.columns);
 
-            switch (getCellType(active, isEditable, dropdown, column.presentation)) {
-                case CellType.Dropdown:
-                    return (<CellDropdown
-                        key={`column-${columnIndex}`}
-                        active={active}
-                        clearable={column.clearable}
-                        dropdown={dropdown}
-                        onChange={handlers.onChange}
-                        value={datum[column.id]}
-                    />);
-                case CellType.Input:
-                    return (<CellInput
-                        key={`column-${columnIndex}`}
-                        active={active}
-                        className={className}
-                        focused={isFocused}
-                        onChange={handlers.onChange}
-                        onClick={handlers.onClick}
-                        onDoubleClick={handlers.onDoubleClick}
-                        onMouseUp={handlers.onMouseUp}
-                        onPaste={handlers.onPaste}
-                        type={column.type}
-                        value={datum[column.id]}
-                    />);
-                case CellType.Label:
-                default:
-                    return (<CellLabel
-                        className={className}
-                        key={`column-${columnIndex}`}
-                        onClick={handlers.onClick}
-                        onDoubleClick={handlers.onDoubleClick}
-                        value={datum[column.id]}
-                    />);
-            }
-        },
-        columns
-    ),
-    data
-);
+                const dropdown = dropdowns[rowIndex][columnIndex];
 
-export default memoizeOneFactory(getter);
+                const isEditable = isCellEditable(editable, column.editable);
+
+                const className = [
+                    ...(active ? ['input-active'] : []),
+                    ...(isFocused ? ['focused'] : ['unfocused']),
+                    ...['dash-cell-value']
+                ].join(' ');
+
+                switch (getCellType(active, isEditable, dropdown, column.presentation)) {
+                    case CellType.Dropdown:
+                        return (<CellDropdown
+                            key={`column-${columnIndex}`}
+                            active={active}
+                            clearable={column.clearable}
+                            dropdown={dropdown}
+                            onChange={this.handlers(Handler.Change, rowIndex, columnIndex)}
+                            value={datum[column.id]}
+                        />);
+                    case CellType.Input:
+                        return (<CellInput
+                            key={`column-${columnIndex}`}
+                            active={active}
+                            className={className}
+                            focused={isFocused}
+                            onChange={this.handlers(Handler.Change, rowIndex, columnIndex)}
+                            onClick={this.handlers(Handler.Click, rowIndex, columnIndex)}
+                            onDoubleClick={this.handlers(Handler.DoubleClick, rowIndex, columnIndex)}
+                            onMouseUp={this.handlers(Handler.MouseUp, rowIndex, columnIndex)}
+                            onPaste={this.handlers(Handler.Paste, rowIndex, columnIndex)}
+                            type={column.type}
+                            value={datum[column.id]}
+                        />);
+                    case CellType.Label:
+                    default:
+                        return (<CellLabel
+                            className={className}
+                            key={`column-${columnIndex}`}
+                            onClick={this.handlers(Handler.Click, rowIndex, columnIndex)}
+                            onDoubleClick={this.handlers(Handler.DoubleClick, rowIndex, columnIndex)}
+                            value={datum[column.id]}
+                        />);
+                }
+            },
+            columns
+        ),
+        data
+    ));
+}
