+++ src/dash-table/derived/cell/eventHandlerProps.ts
@@ -1,39 +1,67 @@
-import {
-    ChangeEvent,
-    ClipboardEvent,
-    MouseEvent
-} from 'react';
+import valueCache from 'core/cache/value';
 
-import { memoizeOneFactory } from 'core/memoizer';
 import { ICellFactoryProps } from 'dash-table/components/Table/props';
-import cellEventHandler, { Handler } from 'dash-table/derived/cell/eventHandler';
+import {
+    handleChange,
+    handleClick,
+    handleDoubleClick,
+    handleEnter,
+    handleLeave,
+    handleMove,
+    handleOnMouseUp,
+    handlePaste
+} from 'dash-table/handlers/cellEvents';
 
-interface IFunctionCache {
-    onChange: (e: ChangeEvent) => void;
-    onClick: (e: MouseEvent) => void;
-    onDoubleClick: (e: MouseEvent) => void;
-    onEnter: (e: MouseEvent) => void;
-    onLeave: (e: MouseEvent) => void;
-    onMove: (e: MouseEvent) => void;
-    onMouseUp: (e: MouseEvent) => void;
-    onPaste: (e: ClipboardEvent<Element>) => void;
+export enum Handler {
+    Change = 'change',
+    Click = 'click',
+    DoubleClick = 'doubleclick',
+    Enter = 'enter',
+    Leave = 'leave',
+    Move = 'move',
+    MouseUp = 'mouseup',
+    Paste = 'paste'
 }
 
-const getter = (propsFn: () => ICellFactoryProps) => {
-    const derivedHandlers = cellEventHandler(propsFn);
+export default (propsFn: () => ICellFactoryProps) => new EventHandler(propsFn).get;
 
-    return (rowIndex: number, columnIndex: number): IFunctionCache => {
-        return {
-            onChange: derivedHandlers(Handler.Change, columnIndex, rowIndex),
-            onClick: derivedHandlers(Handler.Click, columnIndex, rowIndex),
-            onDoubleClick: derivedHandlers(Handler.DoubleClick, columnIndex, rowIndex),
-            onEnter: derivedHandlers(Handler.Enter, columnIndex, rowIndex),
-            onLeave: derivedHandlers(Handler.Leave, columnIndex, rowIndex),
-            onMove: derivedHandlers(Handler.Move, columnIndex, rowIndex),
-            onMouseUp: derivedHandlers(Handler.MouseUp, columnIndex, rowIndex),
-            onPaste: derivedHandlers(Handler.Paste, columnIndex, rowIndex)
-        };
-    };
-};
+class EventHandler {
+    constructor(private readonly propsFn: () => ICellFactoryProps) {
 
-export default memoizeOneFactory(getter);
+    }
+
+    private readonly cache = valueCache<[Handler, number, number]>()((
+        handler: Handler,
+        columnIndex: number,
+        rowIndex: number
+    ) => {
+        switch (handler) {
+            case Handler.Change:
+                return handleChange.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Click:
+                return handleClick.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.DoubleClick:
+                return handleDoubleClick.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Enter:
+                return handleEnter.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Leave:
+                return handleLeave.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Move:
+                return handleMove.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.MouseUp:
+                return handleOnMouseUp.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            case Handler.Paste:
+                return handlePaste.bind(undefined, this.propsFn, rowIndex, columnIndex);
+            default:
+                throw new Error(`unexpected handler ${handler}`);
+        }
+    });
+
+    get = (
+        handler: Handler,
+        columnIndex: number,
+        rowIndex: number
+    ) => {
+        return this.cache.get(handler, columnIndex, rowIndex);
+    }
+}
